fix(config): narrow DistributorCountRequirement int fields
HEU-502. DistributorCountRequirement Go fields were wider than their
Rust mirror, violating the byte-width rule in
docs/development/config-types.md. A Go int accepts an out-of-range value
at YAML unmarshal, then truncates or fails opaquely at the Rust FFI
boundary.

Narrow each Go field to match the verified Rust type in rank.rs:
- Count int -> uint16 (Rust u16)
- TotalCount int -> uint16 (Rust u16)
- SearchDepth *int -> *uint8 (Rust Option<u8>, schema max 255)

Add table-driven boundary tests decoding via yaml.Unmarshal: count and
total_count reject above 65535 and negatives; search_depth rejects above
255 and negatives (0 passes the uint8 type, schema minimum 1 is the gate
at that bound). Update the one test call site to a typed uint8 literal.

+++ internal/config/rules_test.go
@@ -1028,7 +1028,7 @@ func TestValidation_SearchModeFirstLevelsWithoutDepthWarning(t *testing.T) {
 
 func TestValidation_SearchModeFirstLevelsWithDepthPasses(t *testing.T) {
 	plan := minimalPlan()
-	depth := 3
+	depth := uint8(3)
 	plan.Ranks[1].Qualification.Structures = []StructureQualification{
 		{
 			Structure:      "Primary",
